backend/python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
tensorflow==2.15.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
numpy==1.26.3
transformers==4.37.2
openai==1.12.0
python-dotenv==1.0.1
scikit-learn==1.4.0
pytesseract==0.3.10
ultralytics==8.1.11
